refactor: 代码质量优化与架构清理 — 去重/死代码清理/clippy清零
- 提取 encode_png() 到 platform/util.rs，消除 source.rs 与 file_icon.rs 重复代码
- 统一 get_cursor_pos() — 删除 cursor.rs，app.rs 改用 monitor::get_cursor_pos()
- 删除未使用的 BatchToolbar.slint 和 clear-all() 回调
- 删除 is_rgb_format() 死代码（无生产调用）
- 修复 Linux hotkey stub 中 todo!() panic 风险
- 定义 DEFAULT_WINDOW_WIDTH/HEIGHT 常量，消除 5 处硬编码
- 修复全部 17 个 clippy 警告（冗余转换/命名约定/不安全注解等）
- 更新 CLAUDE.md 和 README 至当前项目状态

+++ src/platform/paste.rs
@@ -81,7 +81,7 @@ fn wait_for_focus_and_send_ctrl_v(target_hwnd: Option<usize>) {
         // Ctrl down
         inputs[0].r#type = INPUT_KEYBOARD;
         inputs[0].Anonymous.ki = KEYBDINPUT {
-            wVk: VK_CONTROL as u16,
+            wVk: VK_CONTROL,
             wScan: 0,
             dwFlags: 0,
             time: 0,
@@ -91,7 +91,7 @@ fn wait_for_focus_and_send_ctrl_v(target_hwnd: Option<usize>) {
         // V down
         inputs[1].r#type = INPUT_KEYBOARD;
         inputs[1].Anonymous.ki = KEYBDINPUT {
-            wVk: VK_V as u16,
+            wVk: VK_V,
             wScan: 0,
             dwFlags: 0,
             time: 0,
@@ -101,7 +101,7 @@ fn wait_for_focus_and_send_ctrl_v(target_hwnd: Option<usize>) {
         // V up
         inputs[2].r#type = INPUT_KEYBOARD;
         inputs[2].Anonymous.ki = KEYBDINPUT {
-            wVk: VK_V as u16,
+            wVk: VK_V,
             wScan: 0,
             dwFlags: KEYEVENTF_KEYUP,
             time: 0,
@@ -111,7 +111,7 @@ fn wait_for_focus_and_send_ctrl_v(target_hwnd: Option<usize>) {
         // Ctrl up
         inputs[3].r#type = INPUT_KEYBOARD;
         inputs[3].Anonymous.ki = KEYBDINPUT {
-            wVk: VK_CONTROL as u16,
+            wVk: VK_CONTROL,
             wScan: 0,
             dwFlags: KEYEVENTF_KEYUP,
             time: 0,
